queuelength is now emitted as unsigned long

+++ src/linklayer/shaper/avb/CoRE4INET_AVBShaper.h
@@ -199,7 +199,7 @@ void AVBShaper<SRCLASS, TC>::enqueueMessage(cMessage *msg)
     if (msg->arrivedOn(avbInName.c_str()))
     {
         avbQueue.insert(msg);
-        cComponent::emit(avbQueueLengthSignal, avbQueue.length());
+        cComponent::emit(avbQueueLengthSignal, static_cast<unsigned long>(avbQueue.length()));
         TC::notifyListeners();
     }
     else
@@ -234,7 +234,7 @@ cMessage* AVBShaper<SRCLASS, TC>::pop()
     if (!avbQueue.isEmpty() && avbBuffer->getCredit() >= 0)
     {
         cMessage *msg = static_cast<cMessage*>(avbQueue.pop());
-        cComponent::emit(avbQueueLengthSignal, avbQueue.length());
+        cComponent::emit(avbQueueLengthSignal, static_cast<unsigned long>(avbQueue.length()));
         SimTime duration = TC::outChannel->calculateDuration(msg);
         avbBuffer->sendSlope(duration);
         return msg;
